refactor: 전략, 숫자 생성기 분리
