+++ src/main/kotlin/racingcar/Car.kt
@@ -1,10 +1,13 @@
 package racingcar
 
+import racingcar.movingstrategy.DefaultMovingStrategy
+import racingcar.movingstrategy.MovingStrategy
+
 class Car {
     var position: Int = DEFAULT_POSITION
         private set
 
-    fun move(movingStrategy: MovingStrategy = RandomMovingStrategy()) {
+    fun move(movingStrategy: MovingStrategy = DefaultMovingStrategy()) {
         if (movingStrategy.isMovable()) {
             position++
         }
