+++ src/main/kotlin/racingcar/movingstrategy/MovingStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.movingstrategy
 
 interface MovingStrategy {
     fun isMovable(): Boolean
