+++ src/main/kotlin/racingcar/numbergenerator/NumberGenerator.kt
@@ -1,0 +1,5 @@
+package racingcar.numbergenerator
+
+interface NumberGenerator {
+    fun get(): Int
+}
